Move signal to util
