+++ libopenage/util/signal.h
